# openapi.yaml
openapi: "3.0.3"
info:
  title: Weft API
  description: |
    Crowdsourced Civic Issue Reporting Platform.
    All endpoints are prefixed with `/v1`.
  version: "1.0.0"
  contact:
    name: Weft Engineering
    url: https://weft.city
servers:
  - url: http://localhost:8000/v1
    description: Local development
  - url: https://api.weft.city/v1
    description: Production
...
paths:
...
  /issues:
    post:
      tags: [Issues]
      summary: Submit a new civic issue
      operationId: createIssue
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateIssueResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags: [Issues]
      summary: List issues with filters
      operationId: listIssues
      parameters:
        - name: lat
          in: query
          schema: { type: number, format: double }
        - name: lng
          in: query
          schema: { type: number, format: double }
        - name: radius_m
          in: query
          schema: { type: integer, default: 2000, maximum: 10000 }
        - name: category
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/IssueCategory" } }
        - name: status
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/IssueStatus" } }
        - name: min_upvotes
          in: query
          schema: { type: integer }
        - name: from_date
          in: query
          schema: { type: string, format: date-time }
        - name: to_date
          in: query
          schema: { type: string, format: date-time }
        - name: zone_id
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: sort_by
          in: query
          schema: { type: string, enum: [created_at, upvote_count], default: created_at }
        - name: sort_dir
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Paginated issue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssueResponse"
  /issues/nearby:
    get:
      tags: [Issues]
      summary: Get issues near a location
      operationId: nearbyIssues
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, format: double }
        - name: lng
          in: query
          required: true
          schema: { type: number, format: double }
        - name: radius_m
          in: query
          schema: { type: integer, default: 2000, maximum: 10000 }
      responses:
        "200":
          description: Nearby issues with distance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedIssueResponse"
  /issues/{issue_id}:
    get:
      tags: [Issues]
      summary: Get full issue detail
      operationId: getIssue
      parameters:
        - $ref: "#/components/parameters/IssueId"
      responses:
        "200":
          description: Issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"
        "404":
          description: Issue not found
  /issues/{issue_id}/status:
    patch:
      tags: [Issues]
      summary: Update issue status (authority only)
      operationId: updateIssueStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/IssueStatus"
                note:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDetail"
        "403":
          description: Insufficient permissions
  /issues/{issue_id}/assign:
    patch:
      tags: [Issues]
      summary: Assign issue to field worker (authority only)
      operationId: assignIssue
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assigned_to_id]
              properties:
                assigned_to_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Issue assigned
  /issues/{issue_id}/images:
    post:
      tags: [Issues]
      summary: Upload resolution proof photo (authority only)
      operationId: uploadResolutionProof
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: Image uploaded

  # ── Upvotes ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 signed JWT access token (24h expiry)
  parameters:
    IssueId:
      name: issue_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Issue UUID
  schemas:
...
    IssueStatus:
      type: string
      enum: [REPORTED, IN_PROGRESS, RESOLVED, REJECTED]
    IssueCategory:
      type: string
      enum:
        - POTHOLE
        - GARBAGE_ACCUMULATION
        - WATER_LOGGING
        - BROKEN_STREET_LIGHT
        - DAMAGED_FOOTPATH
        - SEWAGE_OVERFLOW
        - OTHER

    # ── Auth ──────────────────────────────────────────────────────────
...
    CreateIssueRequest:
      type: object
      required: [category, latitude, longitude, image]
      properties:
        category:
          $ref: "#/components/schemas/IssueCategory"
        description:
          type: string
          maxLength: 500
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        image:
          type: string
          format: binary
        captcha_token:
          type: string
          description: Required for anonymous submissions
        captured_at:
          type: string
          format: date-time
          description: When photo was captured (for offline submissions)
    CreateIssueResponse:
      type: object
      properties:
        issue_id:
          type: string
          format: uuid
        issue_number:
          type: string
          example: "ISS-2025-00412"
        status:
          $ref: "#/components/schemas/IssueStatus"
        category:
          $ref: "#/components/schemas/IssueCategory"
        ai_suggested_category:
          $ref: "#/components/schemas/IssueCategory"
        ai_confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        latitude:
          type: number
        longitude:
          type: number
        address_text:
          type: string
        image_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
    IssueDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        issue_number:
          type: string
        category:
          $ref: "#/components/schemas/IssueCategory"
        description:
          type: string
        status:
          $ref: "#/components/schemas/IssueStatus"
        latitude:
          type: number
        longitude:
          type: number
        address_text:
          type: string
        zone_id:
          type: string
          format: uuid
        department_id:
          type: string
          format: uuid
        ai_suggested_category:
          $ref: "#/components/schemas/IssueCategory"
        ai_confidence:
          type: number
        upvote_count:
          type: integer
        assigned_to_id:
          type: string
          format: uuid
          nullable: true
        assigned_at:
          type: string
          format: date-time
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/IssueImage"
    IssueImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cdn_url:
          type: string
          format: uri
        image_type:
          type: string
          enum: [REPORT, RESOLUTION_PROOF]
        created_at:
          type: string
          format: date-time
    PaginatedIssueResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IssueDetail"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # ── Users ─────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          # `details` carries a default of {} and so is not required.
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...